pyproject.toml
[project]
name = "agents-matrix"
version = "0.1.0"
description = "Paid Agent-as-a-Service platform — Calibre ebook operations via A2A + x402"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server]>=0.3.25",
    "x402[fastapi,httpx,evm]>=2.3.0",
    "agent0-sdk>=1.6.0",
    "anthropic>=0.52.0",
    "mcp[cli]>=1.26.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.7.0",
    "tomli>=2.2.0",
    "cli-anything-calibre",
]

[tool.uv]
prerelease = "allow"

[tool.uv.sources]
cli-anything-calibre = { path = "../../calibre/agent-harness", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["config", "server", "executor", "register", "agent", "mcp_server"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
